Fix import errors after refactor
- I will fix all `ModuleNotFoundError` exceptions that were introduced during the refactoring of the file structure.
- The integration will now load correctly.
