+++ custom_components/meraki_ha/config_flow.py
@@ -18,7 +18,7 @@ from homeassistant.exceptions import ConfigEntryAuthFailed
 from homeassistant.helpers import selector
 
 from custom_components.meraki_ha.authentication import validate_meraki_credentials
-from custom_components.meraki_ha.meraki_api import MerakiAPIClient
+from .api.meraki_api import MerakiAPIClient
 from custom_components.meraki_ha.const import (
     CONF_MERAKI_API_KEY,
     CONF_MERAKI_ORG_ID,
