+++ custom_components/meraki_ha/coordinators/data_aggregation_coordinator.py
@@ -154,7 +154,7 @@ class DataAggregationCoordinator(DataUpdateCoordinator[Dict[str, Any]]):
         # _LOGGER.debug(
         #     "DataAggregationCoordinator received raw data. Processing and aggregating..."
         # ) # Reduced verbosity
-        from custom_components.meraki_ha.meraki_api import MerakiApiError
+        from ..api.meraki_api import MerakiApiError
         import aiohttp
 
         try:
