+++ custom_components/meraki_ha/coordinators/tag_eraser.py
@@ -3,7 +3,7 @@
 import logging
 
 from .api_data_fetcher import MerakiApiDataFetcher
-from ..meraki_api.exceptions import (
+from ..api.meraki_api.exceptions import (
     MerakiApiConnectionError,
     MerakiApiAuthError,  # This replaces MerakiApiInvalidApiKeyError
 )
