+++ custom_components/meraki_ha/sensor/device/camera_settings.py
@@ -13,8 +13,8 @@ from homeassistant.core import callback
 from homeassistant.helpers.device_registry import DeviceInfo
 from homeassistant.helpers.update_coordinator import CoordinatorEntity
 
-from ..coordinators import MerakiDataUpdateCoordinator # Ensure this is the correct path
-from ..const import DOMAIN
+from ...coordinators import MerakiDataUpdateCoordinator # Ensure this is the correct path
+from ...const import DOMAIN
 
 _LOGGER = logging.getLogger(__name__)
