+++ custom_components/meraki_ha/sensor/device/device_status.py
@@ -18,8 +18,8 @@ from homeassistant.helpers.device_registry import DeviceInfo
 from homeassistant.helpers.update_coordinator import CoordinatorEntity
 
 # Assuming MerakiDataUpdateCoordinator is the specific coordinator type
-from ..coordinators import MerakiDataUpdateCoordinator
-from ..const import DOMAIN  # For device_info identifiers
+from ...coordinators import MerakiDataUpdateCoordinator
+from ...const import DOMAIN  # For device_info identifiers
 
 _LOGGER = logging.getLogger(__name__)
