+++ custom_components/meraki_ha/sensor/network/meraki_network_info.py
@@ -8,8 +8,8 @@ from homeassistant.core import callback
 from homeassistant.helpers.device_registry import DeviceInfo
 from homeassistant.helpers.update_coordinator import CoordinatorEntity
 
-from ..const import DOMAIN
-from ..coordinators import MerakiDataUpdateCoordinator
+from ...const import DOMAIN
+from ...coordinators import MerakiDataUpdateCoordinator
 
 _LOGGER = logging.getLogger(__name__)
