+++ custom_components/meraki_ha/sensor/network/ssid_availability.py
@@ -17,8 +17,8 @@ from homeassistant.helpers.update_coordinator import CoordinatorEntity
 from homeassistant.helpers.device_registry import DeviceInfo
 
 # Use SSIDDeviceCoordinator for these sensors
-from ..coordinators.ssid_device_coordinator import SSIDDeviceCoordinator
-from ..const import DOMAIN
+from ...coordinators.ssid_device_coordinator import SSIDDeviceCoordinator
+from ...const import DOMAIN
 
 
 _LOGGER = logging.getLogger(__name__)
