+++ custom_components/meraki_ha/sensor/network/ssid_client_count.py
@@ -16,8 +16,8 @@ from homeassistant.helpers.update_coordinator import CoordinatorEntity
 from homeassistant.helpers.device_registry import DeviceInfo
 
 # Use SSIDDeviceCoordinator for these sensors
-from ..coordinators.ssid_device_coordinator import SSIDDeviceCoordinator
-from ..const import DOMAIN
+from ...coordinators.ssid_device_coordinator import SSIDDeviceCoordinator
+from ...const import DOMAIN
 
 _LOGGER = logging.getLogger(__name__)
