+++ custom_components/meraki_ha/sensor/org/org_clients.py
@@ -8,8 +8,8 @@ from homeassistant.core import callback
 from homeassistant.helpers.device_registry import DeviceInfo
 from homeassistant.helpers.update_coordinator import CoordinatorEntity
 
-from ..coordinators import MerakiDataUpdateCoordinator # Adjusted path
-from ..const import DOMAIN
+from ...coordinators import MerakiDataUpdateCoordinator # Adjusted path
+from ...const import DOMAIN
 
 _LOGGER = logging.getLogger(__name__)
